requirements.txt
wxPython
paho-mqtt
pytest
pyside6